third_party/blink/web_tests/flag-specific/highdpi/external/wpt/url/a-element_include=mailto-expected.txt

This is a testharness.js-based test.
[FAIL] Parsing: <mailto:/../> against <about:blank>
  assert_equals: href expected "mailto:/" but got "mailto:/../"
[FAIL] Parsing: <mailto://example.com:8080/pathname?search#hash> against <about:blank>
  assert_equals: href expected "mailto://example.com:8080/pathname?search#hash" but got "mailto://example.com:8080/pathname?search%23hash"
[FAIL] Parsing: <mailto:///test> against <about:blank>
  assert_equals: pathname expected "/test" but got "///test"
[FAIL] Parsing: <mailto://test/a/../b> against <about:blank>
  assert_equals: href expected "mailto://test/b" but got "mailto://test/a/../b"
[FAIL] Parsing: <mailto://:443> against <about:blank>
  assert_unreached: Expected URL to fail parsing Reached unreachable code
[FAIL] Parsing: <mailto://test:test> against <about:blank>
  assert_unreached: Expected URL to fail parsing Reached unreachable code
[FAIL] Parsing: <mailto://[:1]> against <about:blank>
  assert_unreached: Expected URL to fail parsing Reached unreachable code
Harness: the test ran to completion.
